refactor(api): Expose Constraints and StatefulConstraints via api module
Moves all of the core Constraint code into `keel-api`.

Two new interfaces were added to help this effort:

- `com.netflix.spinnaker.keel.api.support.EventPublisher` - A subset interface
  of Spring's `ApplicationEventPublisher`. This is injected into `Constraint`
  classes as `SpringEventPublishedBridge`, which just delegates to Spring.
- `com.netflix.spinnaker.keel.api.constraints.ConstraintRepository` - This is
  just a subset of the god `KeelRepository` interface. Implemented by
  `ConstraintRepositoryBridge`, which just delegates to the `KeelRepository`
  implementation.

This also introduces the `ULID` library as an api dependency of `keel-api`.

I also did some non-functional cleanup: Single class per file rule.
