+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/constraints/ArtifactUsedConstraintEvaluator.kt
@@ -3,18 +3,28 @@ package com.netflix.spinnaker.keel.constraints
 import com.netflix.spinnaker.keel.api.DeliveryConfig
 import com.netflix.spinnaker.keel.api.Environment
 import com.netflix.spinnaker.keel.api.artifacts.DeliveryArtifact
+import com.netflix.spinnaker.keel.api.constraints.ConstraintEvaluator
+import com.netflix.spinnaker.keel.api.constraints.SupportedConstraintType
+import com.netflix.spinnaker.keel.api.support.EventPublisher
 import com.netflix.spinnaker.keel.core.api.ArtifactUsedConstraint
-import org.springframework.context.ApplicationEventPublisher
 import org.springframework.stereotype.Component
 
+/**
+ * TODO: Docs.
+ */
 @Component
 class ArtifactUsedConstraintEvaluator(
-  override val eventPublisher: ApplicationEventPublisher
+  override val eventPublisher: EventPublisher
 ) : ConstraintEvaluator<ArtifactUsedConstraint> {
   override val supportedType = SupportedConstraintType<ArtifactUsedConstraint>("artifact-used")
   override fun isImplicit() = true
 
-  override fun canPromote(artifact: DeliveryArtifact, version: String, deliveryConfig: DeliveryConfig, targetEnvironment: Environment): Boolean {
+  override fun canPromote(
+    artifact: DeliveryArtifact,
+    version: String,
+    deliveryConfig: DeliveryConfig,
+    targetEnvironment: Environment
+  ): Boolean {
     val allowedReferences = mutableSetOf<String>()
     deliveryConfig
       .environments
