+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/constraints/CanaryConstraintDeployHandler.kt
@@ -5,7 +5,13 @@ import com.netflix.spinnaker.keel.api.Environment
 import com.netflix.spinnaker.keel.api.actuation.Task
 import com.netflix.spinnaker.keel.core.api.CanaryConstraint
 
+/**
+ * TODO: Docs.
+ */
 interface CanaryConstraintDeployHandler {
+  /**
+   * TODO: Docs.
+   */
   val supportedClouds: Set<String>
 
   /**
